rename: 컨트롤러 유틸 함수 분리
컨트롤러에서만 사용되는 유틸 함수들을
controller 패키지 내에 util 패키지 하나를 만들어
그곳에 정리함.
